[TYPES] Changing Double into Int64 (long)

+++ LeagueAPI/LeagueAPI/FrameworkAPI/LeagueAPI.swift
@@ -18,19 +18,19 @@ public class LeagueAPI {
     
     // MARK: - Champion Mastery
     
-    public func getChampionMasteries(by summonerId: Double, on region: Region, handler: @escaping ([ChampionMastery]?, String?) -> Void) {
+    public func getChampionMasteries(by summonerId: Int64, on region: Region, handler: @escaping ([ChampionMastery]?, String?) -> Void) {
         ChampionMasteryBusiness.getMastery(method: .BySummonerId(id: summonerId), region: region, key: self.key) { (championMasteries, error) in
             handler(championMasteries, error)
         }
     }
     
-    public func getChampionMastery(by summonerId: Double, for championId: Double, on region: Region, handler: @escaping (ChampionMastery?, String?) -> Void) {
+    public func getChampionMastery(by summonerId: Int64, for championId: Int64, on region: Region, handler: @escaping (ChampionMastery?, String?) -> Void) {
         ChampionMasteryBusiness.getMastery(method: .BySummonerIdAndChampionId(summonerId: summonerId, championId: championId), region: region, key: self.key) { (championMastery, error) in
             handler(championMastery, error)
         }
     }
     
-    public func getMasteryScore(for summonerId: Double, on region: Region, handler: @escaping (Int?, String?) -> Void) {
+    public func getMasteryScore(for summonerId: Int64, on region: Region, handler: @escaping (Int?, String?) -> Void) {
         ChampionMasteryBusiness.getMastery(method: .ScoreBySummonerId(id: summonerId), region: region, key: self.key) { (masteryScore, error) in
             handler(masteryScore, error)
         }
@@ -44,7 +44,7 @@ public class LeagueAPI {
         }
     }
     
-    public func getChampion(by championId: Double, on region: Region, handler: @escaping (Champion?, String?) -> Void) {
+    public func getChampion(by championId: Int64, on region: Region, handler: @escaping (Champion?, String?) -> Void) {
         ChampionBusiness.getChampion(method: .ById(id: championId), region: region, key: self.key) { (champion, error) in
             handler(champion, error)
         }
@@ -72,13 +72,13 @@ public class LeagueAPI {
         }
     }
     
-    public func getRankedPositions(for summonerId: Double, on region: Region, handler: @escaping ([RankedPosition]?, String?) -> Void) {
+    public func getRankedPositions(for summonerId: Int64, on region: Region, handler: @escaping ([RankedPosition]?, String?) -> Void) {
         RankedBusiness.getRanked(method: .PositionsById(id: summonerId), region: region, key: self.key) { (league, error) in
             handler(league, error)
         }
     }
     
-    public func getRankedPosition(for summonerId: Double, in queue: Queue, on region: Region, handler: @escaping (RankedPosition?, String?) -> Void) {
+    public func getRankedPosition(for summonerId: Int64, in queue: Queue, on region: Region, handler: @escaping (RankedPosition?, String?) -> Void) {
         RankedBusiness.getRankedPosition(in: queue, method: .PositionsById(id: summonerId), region: region, key: self.key) { (league, error) in
             handler(league, error)
         }
@@ -94,31 +94,31 @@ public class LeagueAPI {
     
     // MARK: - Match
     
-    public func getMatch(by gameId: Double, on region: Region, handler: @escaping (Match?, String?) -> Void) {
+    public func getMatch(by gameId: Int64, on region: Region, handler: @escaping (Match?, String?) -> Void) {
         MatchBusiness.getMatch(method: .ById(id: gameId), region: region, key: self.key) { (match, error) in
             handler(match, error)
         }
     }
     
-    public func getMatchList(by accountId: Double, on region: Region, beginTime: Double? = nil, endTime: Double? = nil, beginIndex: Int? = nil, endIndex: Int? = nil, championId: Int? = nil, queue: Int? = nil, season: Int? = nil, handler: @escaping (MatchList?, String?) -> Void) {
+    public func getMatchList(by accountId: Int64, on region: Region, beginTime: Int64? = nil, endTime: Int64? = nil, beginIndex: Int? = nil, endIndex: Int? = nil, championId: Int? = nil, queue: Int? = nil, season: Int? = nil, handler: @escaping (MatchList?, String?) -> Void) {
         MatchBusiness.getMatch(method: .MatchesByAccountId(id: accountId, beginTime: beginTime, endTime: endTime, beginIndex: beginIndex, endIndex: endIndex, championId: championId, queue: queue, season: season), region: region, key: self.key) { (match, error) in
             handler(match, error)
         }
     }
     
-    public func getMatchTimeline(by gameId: Double, on region: Region, handler: @escaping (MatchTimeline?, String?) -> Void) {
+    public func getMatchTimeline(by gameId: Int64, on region: Region, handler: @escaping (MatchTimeline?, String?) -> Void) {
         MatchBusiness.getMatch(method: .TimelineById(id: gameId), region: region, key: self.key) { (match, error) in
             handler(match, error)
         }
     }
     
-    public func getMatchIds(by tournamentCode: String, on region: Region, handler: @escaping ([Double]?, String?) -> Void) {
+    public func getMatchIds(by tournamentCode: String, on region: Region, handler: @escaping ([Int64]?, String?) -> Void) {
         MatchBusiness.getMatch(method: .MatchIdsByTournamentCode(code: tournamentCode), region: region, key: self.key) { (match, error) in
             handler(match, error)
         }
     }
     
-    public func getMatch(by gameId: Double, and tournamentCode: String, on region: Region, handler: @escaping (Match?, String?) -> Void) {
+    public func getMatch(by gameId: Int64, and tournamentCode: String, on region: Region, handler: @escaping (Match?, String?) -> Void) {
         MatchBusiness.getMatch(method: .ByIdAndTournamentCode(id: gameId, code: tournamentCode), region: region, key: self.key) { (match, error) in
             handler(match, error)
         }
@@ -126,7 +126,7 @@ public class LeagueAPI {
     
     // MARK: - Spectator
     
-    public func getCurrentGame(by summonerId: Double, on region: Region, handler: @escaping (GameInfo?, String?) -> Void) {
+    public func getCurrentGame(by summonerId: Int64, on region: Region, handler: @escaping (GameInfo?, String?) -> Void) {
         SpectatorBusiness.getCurrentGame(method: .BySummonerId(id: summonerId), region: region, key: self.key) { (game, error) in
             handler(game, error)
         }
@@ -140,7 +140,7 @@ public class LeagueAPI {
     
     // MARK: - Summoner
     
-    public func getSummonerByAccountId(accountId: Double, on region: Region, handler: @escaping (Summoner?, String?) -> Void) {
+    public func getSummonerByAccountId(accountId: Int64, on region: Region, handler: @escaping (Summoner?, String?) -> Void) {
         SummonerBusiness.getSummoner(method: .ByAccountId(id: accountId), region: region, key: self.key) { (summoner, error) in
             handler(summoner, error)
         }
@@ -153,7 +153,7 @@ public class LeagueAPI {
     }
     
     // WARNING: Too many calls to unexisting league may result in Blacklist
-    public func getSummoner(by summonerId: Double, on region: Region, handler: @escaping (Summoner?, String?) -> Void) {
+    public func getSummoner(by summonerId: Int64, on region: Region, handler: @escaping (Summoner?, String?) -> Void) {
         SummonerBusiness.getSummoner(method: .ById(id: summonerId), region: region, key: self.key) { (summoner, error) in
             handler(summoner, error)
         }
@@ -161,7 +161,7 @@ public class LeagueAPI {
     
     // MARK: - Third Party Code
     
-    public func getThirdPartyVerificationCode(by summonerId: Double, on region: Region, handler: @escaping (String?, String?) -> Void) {
+    public func getThirdPartyVerificationCode(by summonerId: Int64, on region: Region, handler: @escaping (String?, String?) -> Void) {
         ThirdPartyCodeBusiness.getVerificationCode(method: .ById(id: summonerId), region: region, key: self.key) { (code, error) in
             handler(code, error)
         }
@@ -174,14 +174,14 @@ public class LeagueAPI {
             guard let providerId = providerId, error == nil else { handler(nil, error); return }
             self.createTournament(providerId: providerId, named: name, on: region) { (tournamentId, error) in
                 guard let tournamentId = tournamentId, error == nil else { handler(nil, error); return }
-                self.createTournamentCode(tournamentId: Double(tournamentId), on: region, amount: amount, info: info) { (codes, error) in
+                self.createTournamentCode(tournamentId: Int64(tournamentId), on: region, amount: amount, info: info) { (codes, error) in
                     handler((providerId, tournamentId, codes ?? []), error)
                 }
             }
         }
     }
     
-    public func createTournamentCode(tournamentId: Double, on region: Region, amount: Int? = nil, info: TournamentInfo, handler: @escaping ([String]?, String?) -> Void) {
+    public func createTournamentCode(tournamentId: Int64, on region: Region, amount: Int? = nil, info: TournamentInfo, handler: @escaping ([String]?, String?) -> Void) {
         TournamentStubBusiness.manageTournament(method: .CreateCodes(amount: amount, tournamentId: tournamentId, info: info), region: region, key: self.key) { (codes, error) in
             handler(codes, error)
         }
