+++ LeagueAPI/LeagueAPI/Models/RiotModels/ChampionMastery.swift
@@ -10,15 +10,15 @@ import Foundation
 
 public class ChampionMastery: Decodable {
     
-    public var championId: Double
-    public var summonerId: Double
+    public var championId: Int64
+    public var summonerId: Int64
     public var championLevel: Int
     public var championPoints: Int
-    public var championPointsUntilNextLevel: Double
-    public var championPointsSinceLastLevel: Double
+    public var championPointsUntilNextLevel: Int64
+    public var championPointsSinceLastLevel: Int64
     public var chestGranted: Bool
     public var tokensEarned: Int
-    public var lastPlayTime: Double
+    public var lastPlayTime: Int64
     
     enum CodingKeys: String, CodingKey {
         case championId = "championId"
@@ -32,7 +32,7 @@ public class ChampionMastery: Decodable {
         case lastPlayTime = "lastPlayTime"
     }
     
-    public init(championId: Double, summonerId: Double, championLevel: Int, championPoints: Int, championPointsUntilNextLevel: Double, championPointsSinceLastLevel: Double, chestGranted: Bool, tokensEarned: Int, lastPlayTime: Double) {
+    public init(championId: Int64, summonerId: Int64, championLevel: Int, championPoints: Int, championPointsUntilNextLevel: Int64, championPointsSinceLastLevel: Int64, chestGranted: Bool, tokensEarned: Int, lastPlayTime: Int64) {
         self.championId = championId
         self.summonerId = summonerId
         self.championLevel = championLevel
@@ -46,14 +46,14 @@ public class ChampionMastery: Decodable {
     
     public required init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.championId = try container.decode(Double.self, forKey: .championId)
-        self.summonerId = try container.decode(Double.self, forKey: .summonerId)
+        self.championId = try container.decode(Int64.self, forKey: .championId)
+        self.summonerId = try container.decode(Int64.self, forKey: .summonerId)
         self.championLevel = try container.decode(Int.self, forKey: .championLevel)
         self.championPoints = try container.decode(Int.self, forKey: .championPoints)
-        self.championPointsUntilNextLevel = try container.decode(Double.self, forKey: .championPointsUntilNextLevel)
-        self.championPointsSinceLastLevel = try container.decode(Double.self, forKey: .championPointsSinceLastLevel)
+        self.championPointsUntilNextLevel = try container.decode(Int64.self, forKey: .championPointsUntilNextLevel)
+        self.championPointsSinceLastLevel = try container.decode(Int64.self, forKey: .championPointsSinceLastLevel)
         self.chestGranted = try container.decode(Bool.self, forKey: .chestGranted)
         self.tokensEarned = try container.decode(Int.self, forKey: .tokensEarned)
-        self.lastPlayTime = try container.decode(Double.self, forKey: .lastPlayTime)
+        self.lastPlayTime = try container.decode(Int64.self, forKey: .lastPlayTime)
     }
 }
